requirements.txt
# This is a generated file; DO NOT EDIT!
#
# Please edit 'setup.cfg' to add top-level dependencies and use
# './contrib/refresh-requirements.sh to regenerate this file

amqp==2.6.0
attrs==19.3.0
backports.csv==1.0.7
billiard==3.6.3.0
celery==4.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.29
django-allauth==0.35.0
django-appconf==1.0.4
django-autoslug==1.9.7
django-avatar==4.1.0
django-braces==1.12.0
django-compressor==2.2
django-crispy-forms==1.7.2
django-csp==3.4
django-debug-toolbar==1.11
django-filter==1.1.0
django-jsonfield==1.0.1
django-redis-cache==2.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.7.7
dnspython==1.16.0
docopt==0.6.2
ephem==3.7.7.1
eventlet==0.25.2
greenlet==0.4.16
gunicorn==19.9.0
idna==2.10
internetarchive==1.9.4
jsonpatch==1.26
jsonpointer==2.0
kombu==4.6.11
Logbook==1.5.3
Markdown==2.6.11
monotonic==1.5
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==5.0.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.1
python-dotenv==0.6.5
python3-openid==3.2.0
pytz==2020.1
rcssmin==1.0.6
redis==3.5.3
Represent==1.6.0
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.0.12
rush==2018.12.1
satellitetle==0.9.0
schema==0.7.2
sentry-sdk==0.14.4
sgp4==2.12
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==1.2.0
social-auth-core==3.3.3
spacetrack==0.14.0
sqlparse==0.3.1
tinytag==1.3.1
total-ordering==0.1.0
tqdm==4.47.0
unicode-slugify==0.1.3
Unidecode==1.1.1
Unipath==1.1
urllib3==1.25.9
vine==1.3.0
